+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 
 ### General improvements
 
-- Extended test suite for the replay_checkups command.
+- Extended test suite for the `replay_checkups` command.
+
+### Bugfixes
+
+- The `quickstart_dev` management command no longer tries to create the same organization three
+  times in the report module -- only once.
 
 ## Version 3.29.2, 31st July 2025